first swing at implementing a GUIController, still very much broken
